itadmin_agent/requirements.txt
# Core dependencies
langchain>=0.0.267
openai>=1.3.7
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.4.2
python-dotenv>=1.0.0

# Storage & Database
faiss-cpu>=1.7.4
chromadb>=0.4.18
sqlalchemy>=2.0.23
alembic>=1.12.1

# Remote execution
paramiko>=3.3.1
fabric>=3.2.2
ansible-runner>=2.3.4

# Security
cryptography>=41.0.5
pyjwt>=2.8.0
passlib>=1.7.4
bcrypt>=4.0.1

# Monitoring
prometheus-client>=0.17.1
pydantic-settings>=2.0.3

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0

# UI
streamlit>=1.28.2
plotly>=5.18.0

# Utilities
pyyaml>=6.0.1
jinja2>=3.1.2
python-dateutil>=2.8.2 